move from_matched and compare to new module
